workspace: audit round — sound unsafe, durable stores, off-thread lipc, true hit rects
Safety/soundness:
- input.rs: evdev events read via read_unaligned from the byte buffer
  (casting [u8;N] to &[InputEvent] was misaligned-reference UB);
  guard.rs: TERM/INT handler reduced to an atomic store, restore+exit
  runs from the app loop via with_quit_check (no malloc/fs in the
  signal frame); panel.rs blit in usize (u32 w*h wrapped);
  sysinfo.rs Ifr size assert; SendDoc carries its SAFETY argument

Untrusted input:
- fb2 zip path through read_capped (the in-memory path enforced the
  cap, the disk path — the one the reader uses — did not);
  <binary> base64 accumulation capped at 8 MB; epub spine items that
  are missing/non-UTF-8 land a placeholder chapter (TOC indices stay
  stable); img.rs rejects APNG first-frames smaller than IHDR (1 KB
  remote DoS panic, regression test added); vocab.bin count clamped
  to the index region (armv7 mid*20 overflow); receive: duplicate
  conflicting Content-Length → 400, error bodies escaped, body write
  capped to the declared length (u64 underflow bypass), SYSTEM_FILES
  protected on move/delete too

Durability: notes/positions/flashcards writes route through
ybdev::atomic::write with failure plogs (notes was tmp+rename with
no fsync and both errors swallowed)

UI-thread discipline: system screen snapshots wifi state + screensaver
scan on entry instead of per-draw lipc round-trips; curtain SSH-card
and awake wifi-heal run off-thread with one-at-a-time latches;
ai_stream: named thread per concern, poller joined on Drop, tick 2s
when idle

Correctness: ai_stream sheet + crop dialog row-2 hit regions derived
from the same geometry draw paints (mis-taps on neighbours gone);
toc/footnote/word dialogs truncate() fed points, not pixels (4.2x
too-wide ellipsizing); sel_save bounds-guarded and re-renders clear
the pending selection; yread sub packed via pack_yread_sub (chapters
>1M chars clamp instead of restoring into the wrong chapter);
cache keys hashed with fnv1a (name-sanitized collisions);

wifi: manual-off latch persisted + hydrated at boot, outranks wake
intents (field 2026-08-23); receive PAGE extracted to
receive_page.html (include_str!), one thread per connection;
clippy sweep (~90 → ~55, remaining are structural deferrals);
release profile: overflow-checks, thin LTO, 1 CGU, strip — panic
stays unwind for the catch_unwind parsers; toolchain pinned 1.96.0

+++ reader/src/scrubber_dialog.rs
@@ -56,7 +56,7 @@ fn step_rects(card: Rect) -> Vec<Rect> {
     let h = pt(32.0);
     let w = (card.w - pt(48.0) - pt(24.0)) / 4;
     (0..4)
-        .map(|i| Rect::new(card.x + pt(24.0) + i as i32 * (w + pt(8.0)), y, w, h))
+        .map(|i| Rect::new(card.x + pt(24.0) + i * (w + pt(8.0)), y, w, h))
         .collect()
 }
